Fix __main__ to only delete a directory if that directory actually exists

+++ JackAnalyzer/__main__.py
@@ -15,10 +15,12 @@ if __name__ == '__main__':
     sources = [source]
 
   root_dir = Path(source.stem)
-  for item in root_dir.iterdir():
-    item.unlink()
+  if root_dir.is_dir():
+    for item in root_dir.iterdir():
+      item.unlink()
 
-  root_dir.rmdir()
+    root_dir.rmdir()
+
   root_dir.mkdir()
   for child in sources:
     if child.suffix != '.jack':
